requirements.txt
tavily-python
langgraph
langchain
langchain-openai
crawl4ai
pydantic
asyncio 